Simplify the numbers remaining list
We use a linked list to store the numbers remaining to be drawn. This
starts off containing 1..90

Rather than storing a structure that contains the number as a member,
store the number directly by casting it to a (void *)(long), essentially
storing it as a pointer value. To get the numbers back out we just cast
them like (u8)(long) as u8 is what the variable that stores the return
value from ac_slist_nth_data() is defined as.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>